made enemy team buildings not selectable
